requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.14
deprecation==2.1.0
Flask==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
grpcio==1.56.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.251
langsmith==0.0.18
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
peewee==3.16.2
protobuf==4.23.4
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tinkoff==0.1.1
tinkoff-investments==0.2.0b59
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
